- refactored dio factory - refactored api error model & error handler
