+++ lib/core/networking/api_constants.dart
@@ -1,5 +1,5 @@
 class ApiConstants {
-  static const String baseUrl = '';
+  static const String baseUrl = 'http://connectchain.runasp.net/api/';
 
   static const Duration timeOut = Duration(seconds: 20);
 }
